removed lvm2 and splitted produce. We need now to rebuild LVM2
