+++ src/krill/classes/sequence.d/network-cfg.ts
