+++ src/krill/classes/sequence.d/partition.d/bios-standard.ts
