+++ src/krill/classes/sequence.d/partition.d/manual.ts
