+++ src/krill/classes/sequence.d/partition.d/uefi-standard.ts
