+++ src/krill/classes/sequence.d/remove-installer-link.ts
