+++ src/krill/components/summary.tsx
@@ -16,6 +16,7 @@ import { ISettings, IBranding } from '../../interfaces/index.js'
 import { Text, Box, Newline} from 'ink'
 
 
+
 type SummaryProps = {
 
   language?: string,
